More reorg and formatting
