+++ GUI/Dialogs/InstalledDevicesDlg.py
@@ -35,10 +35,10 @@ class InstalledDevicesDlg(wx.Dialog):
             #                 self.appNameList.append(key)
             #                 break
             # else:
-                # if app["app_name"] not in self.appNameList:
-                #     self.appNameList.append(app["app_name"])
-                # elif "appPkgName" in app:
-                self.appNameList.append(app["appPkgName"])
+            # if app["app_name"] not in self.appNameList:
+            #     self.appNameList.append(app["app_name"])
+            # elif "appPkgName" in app:
+            self.appNameList.append(app["appPkgName"])
         self.versions = []
 
         self.SetMinSize((400, 300))
@@ -49,7 +49,7 @@ class InstalledDevicesDlg(wx.Dialog):
         self.panel_1 = wx.Panel(self, wx.ID_ANY)
         sizer_1.Add(self.panel_1, 1, wx.ALL | wx.EXPAND, 5)
 
-        grid_sizer_1 = wx.GridSizer(1,2,0,0)
+        grid_sizer_1 = wx.GridSizer(1, 2, 0, 0)
 
         grid_sizer_3 = wx.FlexGridSizer(3, 1, 0, 0)
         grid_sizer_1.Add(grid_sizer_3, 1, wx.EXPAND, 0)
