+++ Utility/GroupUtility.py
@@ -4,11 +4,19 @@ import json
 import time
 
 from esperclient.rest import ApiException
+from Common.decorator import api_tool_decorator
+from Utility import EventUtility
 from Utility.ApiToolLogging import ApiToolLog
 import esperclient
 import Common.Globals as Globals
 
-from Utility.Resource import getHeader, performPatchRequestWithRetry
+from Utility.Resource import (
+    getHeader,
+    logBadResponse,
+    performGetRequestWithRetry,
+    performPatchRequestWithRetry,
+    postEventToFrame,
+)
 
 
 def moveGroup(groupId, deviceList, maxAttempt=Globals.MAX_RETRY):
@@ -41,7 +49,7 @@ def createGroup(groupName, groupParent, maxAttempt=Globals.MAX_RETRY):
             try:
                 api_response = api_instance.create_group(Globals.enterprise_id, data)
                 ApiToolLog().LogApiRequestOccurrence(
-                    "getAllGroups", api_instance.get_all_groups, Globals.PRINT_API_LOGS
+                    "create_group", api_instance.create_group, Globals.PRINT_API_LOGS
                 )
                 break
             except Exception as e:
@@ -88,3 +96,114 @@ def renameGroup(groupId, newName):
     body = {"name": newName}
     resp = performPatchRequestWithRetry(url, headers=getHeader(), json=body)
     return resp
+
+
+@api_tool_decorator()
+def fetchGroupName(groupURL, returnJson=False):
+    headers = getHeader()
+    resp = performGetRequestWithRetry(groupURL, headers=headers)
+    try:
+        if resp and resp.status_code < 300:
+            json_resp = resp.json()
+            logBadResponse(groupURL, resp, json_resp)
+
+            if "name" in json_resp:
+                if returnJson:
+                    return json_resp
+                else:
+                    return json_resp["name"]
+    except Exception as e:
+        ApiToolLog().LogError(e)
+        logBadResponse(groupURL, resp, None)
+    return None
+
+
+@api_tool_decorator()
+def getAllGroups(name="", limit=None, offset=None, maxAttempt=Globals.MAX_RETRY):
+    """ Make a API call to get all Groups belonging to the Enterprise """
+    if not limit:
+        limit = Globals.limit
+    if not offset:
+        offset = Globals.offset
+    try:
+        api_instance = esperclient.DeviceGroupApi(
+            esperclient.ApiClient(Globals.configuration)
+        )
+        api_response = None
+        for attempt in range(maxAttempt):
+            try:
+                api_response = api_instance.get_all_groups(
+                    Globals.enterprise_id,
+                    name=name,
+                    limit=limit,
+                    offset=offset,
+                )
+                ApiToolLog().LogApiRequestOccurrence(
+                    "getAllGroups", api_instance.get_all_groups, Globals.PRINT_API_LOGS
+                )
+                break
+            except Exception as e:
+                if attempt == maxAttempt - 1:
+                    ApiToolLog().LogError(e)
+                    raise e
+                time.sleep(Globals.RETRY_SLEEP)
+        postEventToFrame(EventUtility.myEVT_LOG, "---> Group API Request Finished")
+        return api_response
+    except ApiException as e:
+        raise Exception(
+            "Exception when calling DeviceGroupApi->get_all_groups: %s\n" % e
+        )
+
+
+@api_tool_decorator()
+def getDeviceGroupsForHost(config, enterprise_id, maxAttempt=Globals.MAX_RETRY):
+    try:
+        api_instance = esperclient.DeviceGroupApi(esperclient.ApiClient(config))
+        api_response = None
+        for attempt in range(maxAttempt):
+            try:
+                api_response = api_instance.get_all_groups(
+                    enterprise_id, limit=Globals.limit, offset=Globals.offset
+                )
+                ApiToolLog().LogApiRequestOccurrence(
+                    "getDeviceGroupsForHost",
+                    api_instance.get_all_groups,
+                    Globals.PRINT_API_LOGS,
+                )
+                break
+            except Exception as e:
+                if attempt == maxAttempt - 1:
+                    ApiToolLog().LogError(e)
+                    raise e
+                time.sleep(Globals.RETRY_SLEEP)
+        return api_response
+    except Exception as e:
+        raise e
+
+
+@api_tool_decorator()
+def createDeviceGroupForHost(
+    config, enterprise_id, group, maxAttempt=Globals.MAX_RETRY
+):
+    try:
+        api_instance = esperclient.DeviceGroupApi(esperclient.ApiClient(config))
+        api_response = None
+        for attempt in range(maxAttempt):
+            try:
+                api_response = api_instance.create_group(
+                    enterprise_id, data={"name": group}
+                )
+                ApiToolLog().LogApiRequestOccurrence(
+                    "createDeviceGroupForHost",
+                    api_instance.create_group,
+                    Globals.PRINT_API_LOGS,
+                )
+                break
+            except Exception as e:
+                if attempt == maxAttempt - 1:
+                    ApiToolLog().LogError(e)
+                    raise e
+                time.sleep(Globals.RETRY_SLEEP)
+        return api_response
+    except Exception as e:
+        raise e
